refactor: extract controls panel component
